fix: :bug: add numeric to price in order, product and bundle

+++ src/bundle/infraestructure/entities/orm-entities/orm-bundle-entity.ts
@@ -13,7 +13,7 @@ export class OrmBundleEntity implements IBundle{
     @Column( 'timestamp', { default: () => 'CURRENT_TIMESTAMP' } ) caducityDate: Date
     @OneToMany( () => OrmBundleImage,   image => image.bundle,{ eager: true }) images: OrmBundleImage[]   
     @Column( 'integer' ) stock: number
-    @Column( 'integer' ) price: number
+    @Column('numeric') price: number
     @Column( 'varchar' ) currency: string;
     @Column( 'integer' ) weigth: number;
     @Column( 'varchar' ) measurament: string;
